Move already extracted helper modules/classes over to their own support files, instead of polluting helper namespace
