+++ activerecord/test/cases/dirty_test.rb
@@ -1,4 +1,5 @@
 require 'cases/helper'
+require 'support/in_time_zone'
 require 'models/topic'    # For booleans
 require 'models/pirate'   # For timestamps
 require 'models/parrot'
