+++ activerecord/test/support/in_time_zone.rb
@@ -1,0 +1,15 @@
+module InTimeZone
+  private
+
+  def in_time_zone(zone)
+    old_zone = Time.zone
+    old_tz = ActiveRecord::Base.time_zone_aware_attributes
+
+    Time.zone = zone ? ActiveSupport::TimeZone[zone] : nil
+    ActiveRecord::Base.time_zone_aware_attributes = !zone.nil?
+    yield
+  ensure
+    Time.zone = old_zone
+    ActiveRecord::Base.time_zone_aware_attributes = old_tz
+  end
+end
