+++ activerecord/test/support/sql_subscriber.rb
@@ -1,0 +1,18 @@
+class SQLSubscriber
+  attr_reader :logged
+  attr_reader :payloads
+
+  def initialize
+    @logged = []
+    @payloads = []
+  end
+
+  def start(name, id, payload)
+    @payloads << payload
+    @logged << [payload[:sql].squish, payload[:name], payload[:binds]]
+  end
+
+  def finish(name, id, payload)
+    ;
+  end
+end
